have to move it down to the store where I have access to DB pool
